Move xml-to-yaml converter to obsolete
