+++ obsolete/tools/js/convertAllTestsuites.js
