+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.0.0",
   "description": "OASIS OData TC: ABNF for OData URLs, headers, and literal data values",
   "scripts": {
-    "convert-all-testsuites": "node lib/convertAllTestsuites.js",
     "generate-grammar": "node node_modules/apg-js/bin/apg.sh --strict -i ./abnf/odata-abnf-construction-rules.txt,./abnf/odata-aggregation-abnf.txt,./abnf/odata-temporal-abnf.txt -o lib/grammar.js",
     "run-testcases": "node lib/checkAbnf.js",
     "test": "npm run generate-grammar && node lib/checkAbnf.js",
@@ -27,7 +26,6 @@
   "dependencies": {
     "apg-js": "^4.1.1",
     "colors": "^1.4.0",
-    "sax": "^1.2.4",
     "yaml": "^1.10.2"
   },
   "devDependencies": {
